refactor: 리뷰 반영
- svg파일 assets/icons 폴더로 이동
- 삼항연산자 -> &&로 문법 변경
- svg 이미지 컴포넌트화

resolved:#37
