+++ src/components/TeamListItem/common/HashTag/HashTag.js
@@ -1,20 +1,10 @@
 import React from 'react';
 import { HashTagStyled } from './HashTag.styles';
 
-const HashTags = ({ tags}) => {
-
-	const hashTagsLoop = (loopLen) => {
-		const hashTagArr = [];
-		for (let i = 0; i < loopLen; i++)
-			tags[i] && hashTagArr.push(<HashTagStyled key={i}># {tags[i]}</HashTagStyled>);
-		return hashTagArr;
-	}
-
-	return (
-		<span>
-			{hashTagsLoop(5)}
-  	</span>
-	)
-};
+const HashTags = ({ tags }) => (
+	<span>
+		{tags.map((tag, index) => <HashTagStyled key={index}># {tag}</HashTagStyled>)}
+  </span>
+)
 
 export default HashTags;
