+++ src/components/TeamMemberView/TeamMemberImage/TeamMemberImage.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import { UserImageStyled } from '../../TeamListItem/common/TeamImage/TeamImage.styles';
 import TeamMemberImageStyled from './TeamMemberImageStyled';
-import Crown from '../../../assets/images/TeamMemberView_Crown.svg';
-import Empty from '../../../assets/images/TeamMemberView_Empty.svg';
+import { ReactComponent as CrownIcon } from '../../../assets/icons/TeamMemberView_Crown.svg';
+import { ReactComponent as EmptyIcon } from '../../../assets/icons/TeamMemberView_Empty.svg';
 
 const TeamMemberImage = ({ teamMember }) => {
 
 	const teamMemberImageArray = (teamMember) => (
 		teamMember.map((member, index) => (
 			<TeamMemberImageStyled.Item key={index}>
-				{index === 0 ? <img src={Crown} alt="leaderMark" width="16" /> : null}
+				{index === 0 && <CrownIcon />}
 				<UserImageStyled size="big" url={`https://cdn.intra.42.fr/users/small_${member}.jpg`} />
 				{member}
 			</TeamMemberImageStyled.Item>
@@ -23,7 +23,7 @@ const TeamMemberImage = ({ teamMember }) => {
 		for (let i = 0; i < len; i++) {
 			Array.push(
 				<TeamMemberImageStyled.Item key={i + 3}>
-					<UserImageStyled size="big" src={Empty} />
+					<EmptyIcon width="53px" height="53px" />
 					<TeamMemberImageStyled.EmptyName>(empty)</TeamMemberImageStyled.EmptyName>
 				</TeamMemberImageStyled.Item>
 			)
